refactored and restructured server. server is up and running
