+++ .gitignore
@@ -1,4 +1,5 @@
 .DS_Store
+**/.DS_Store
 
 # Files that might appear on external disk
 .Spotlight-V100
