--- server/db/trade.js
@@ -1,12 +1,0 @@
-var mongoose = require('mongoose');
-
-var db = require('./config');
-
-var tradeSchema = mongoose.schema({
-  mealOneId: Number,
-  mealTwoId: Number
-});
-
-var Trade = mongoose.model('Trade', tradeSchema);
-
-module.exports = Trade;
