+++ server/users/user.js
@@ -1,0 +1,10 @@
+var mongoose = require('mongoose');
+
+var userSchema = mongoose.Schema({
+  username: String,
+  password: String
+});
+
+var User = mongoose.model('User', userSchema);
+
+module.exports = User;
